Implement history based navigation
This CL introduces 2 kinds of history:
1. nudge history. When nudging back, it will focus an element within
the previously foucsed FocusArea (if any) rather than the geometrically
close FocusArea.
2. focus history within a FocusArea. When nudging back, it will focus
the previously focused element (if any) rather than the geometrically
close element.

This CL also removes the depth limit when searching focus areas.

TODO: In this CL navigation history has the highest priority. We may
want to change it in the future.

Fixes: 151349253
Test: atest NavigationHelperTest.java && atest RotaryCacheTest.java
Change-Id: Ib993ebac7facdaf5b213574823fd04d4b3f83da3