Handle FocusParkingView
When we need to focus a node in another window, we should "park"
the focus on the FocusParkingView in the current  window. This way
we can fix the auto-focus issue.

Other than that, FocusParkingView shouldn't take focus. So we need
to update our focus finding logic. There are 3 cases:
Case 1: find the first node to focus. We should skip the
FocusParkingViews and keep searching until we find a regular view
Case 2: find rotaton target. Nothing needs to be done because
FocusParkingViews are outside of focus areas, and we'll stop
searching when the target is not in the same focus area with the
source node.
Case 3: find nudge target. We shouldn't count FocusParkingViews as
candidates.

Besides, we should ignore TYPE_VIEW_FOCUSED events for
FocusParkingView.

Bug: 154361882
Test: manual
atest NavigatorTest.java

Change-Id: I0b560e7e79aee7971c24ef3874d23ba11da4e0e3