Support nudging directly to HUN and back.
The heads-up notification (HUN) window is treated specially. Nudging
in the direction of the HUN jumps straight to it and nudging back
returns to where you were. When the HUN closes, focus automatically
returns to where you were. The HUN is identified by its bounds. A new
type of history is introduced to restore the focus when the HUN
disappears.

Additional minor changes:

1. Stop when you hit a FocusParkingView while rotating. This prevents
   rotation from wrapping around when no FocusAreas are declared.

2. Ignore TYPE_VIEW_CLICKED events resulting from injected
   KEYCODE_DPAD_CENTER events in addition to events resulting from
   ACTION_CLICK. We don’t know how the application will handle
   KEYCODE_DPAD_CENTER events so we ignore TYPE_VIEW_CLICKED events
   for a limited period of time.

3. Clear focus before setting the last touched node since setting the
   last touched node clears the focused node.

4. Handle the fail-safe case where the target node is already focused
   better.

Test: atest CarRotaryControllerRoboTests
Bug: 155681517
Change-Id: I3d2eb7ba3bd63178e5d7d2f6898d49bc25586bed