Handle TYPE_VIEW_FOCUSED event properly
3 cases when the FocusParkingView is focused:

Case 1. We focused the FocusParkingView to clear focus. Nothing
needs to be done.
Case 2. We injected a MotionEvent to scroll a scrollable container
and the focused view was scrolled out of the viewport. Android
focused the FocusParkingView automatically. We should move focus to
the scroll container.
Case 3. Other cases, such as:
  a. the focused view was removed, then Android focused the
  FocusParkingView automatically
  b. a dialog window was opened, then Android focused the first
  focusable view (the FocusParkingView) in the dialog window
  automatically.
We should move focus to a view nearby.

Bug: 158797952
Test: manual
Change-Id: I3a05c4f7757efd556c1643daec6ee0055d0af9b0