Update direct manipulaiton handling logic
Previously the only way for a view in application window to enter
direct manipulation mode is to register key event listener and handle
it on app side. This reqiures a lot of work on app developers
side, especially when the view only supports rotating. This CL allows
a view in application window to enter direct manipulation directly
as long as it has specified that. As a result, we don't need to mock
the system window any more.

Bug: 161483857
Test: manual

Change-Id: I54c77911e8d7e040fc025ad0e70708f0621d4674