Validate IME component name
The default touch IME can be configured by the
default_touch_input_method resource.

In this change, RotaryService will validate the IME component name. If
the IME is not found or enabled in the system, RotaryService will not
switch between the two IMEs because it won't be able to switch back.

Bug: 167259679
Bug: 168562905
Test: manual
Change-Id: I09f10e9ce2e592206d27d60a19a18fa267bdab86