Refactor RotaryService and Navigator
1. Update handleViewFocusedEvent(). The focus will be adjusted properly
by the framework due to a recent change, so RotaryService doesn't need
to adjust the focus any more. As a result, PendingFocusedNodes class is
not needed any more.
2. Remove RotaryCache since the history is kept in FocusArea class due
to a recent change.
3. Due to a recent change, the nudge target finding is mostly done on
the app side. So update the logic in RotaryService accordingly.
4. Remove failing roboletric tests. Since roboletric tests are being
deprecated, don't bother updating them to make them pass. Equivalent
instrumented tests will be added later.
5. Replace NodePredicate with Predicate in Java library.

Bug: 168844996
Bug: 170344916
Test: atest CarRotaryControllerRoboTests, and manual test
Change-Id: Ib3199df71da625f5dd907f23bc6b8625a121ca36