Handle fast rotation and nudging
When the user rotates quickly, the TYPE_VIEW_FOCUSED events for the
focus changes can arrive after the focus has already moved. If the event
is for a view which has since scrolled off the screen, its source will
be null. In order to handle fast rotation, RotaryService must update
mFocusedNode when performing ACTION_FOCUS rather than relying on the
subsequent TYPE_VIEW_FOCUSED event.

When performing an action that changes the focus in a way which can't be
predicted ahead of time, e.g. performing ACTION_FOCUS on a FocusArea,
find the focus immediately afterward rather than waiting for the
TYPE_VIEW_FOCUSED event.

In order to determine whether mFocusedNode should be updated for each
TYPE_VIEW_FOCUSED events, keep track of the last node focused by the
RotaryService, whether by performing ACTION_FOCUS on it or by performing
an action which changes the focus unpredictably as described above,
until the corresponding event is received. This is sufficient to
distinguish between externally triggered events, which should set
mFocusedNode, and self-triggered events, which shouldn't.

Test: rotate quickly with occasional nudges on emulator and Seahawk
Bug: 169886834
Change-Id: Ic35bbb448e315de9b5e9e4b0ddcf8be2bcb44b30