Ensure that the focused node isn't reset
This CL removes some legacy code which resets the focused node in
initFocus. This is no longer required since we restore focus in many
other places.

This is needed to resolve an issue with direct manipulation. When
dealing with complex components (i.e. DatePicker/TimePicker) calling
this piece of code causes us to prematurely leave the direct
manipulation mode.

Bug: 169884295
Test: manual (RotaryPlayground + Settings)
Change-Id: I23b1554d2110ff3f1ff7427c88731a7f424d8886