Focus HUN with FPV instead of the FA
This allows us to use FAs for each notification card (so that rotation
stays within each card). The FPV will still focus a view in the HUN
window based on ViewUtils#adjustFocus, so we're able to focus the
appropriate view.

This approach also deals with the stray focus that we have when we see
the first HUN in rotary mode.

Bug: 172267951
Bug: 172268552
Test: manual
Change-Id: I8a3af5ec60749f8463d6655b232d9c67da083c6e