Ignore obscured views when nudging
We require the target focus area to contain a focusable view that's
in the direction of the nudge, but we don't require this view to be
visible to the user. I accidentally removed this requirement when I
made it possible to rotate to a view that's scrolled off the screen
in ag/12317572. This change brings back the requirement for nudging
without affecting rotation.

Test: nudge down in Rotary IME in Cards tab of Rotary Playground
Test: rotate to off-screen view in Grid tab of Rotary Playground
Bug: 177330519
Change-Id: I3028037b662d55147918dab7a75d6e055cedd167