Add a unit test for Navigator
This CL adds a unit test to verify ag/12913229.
This CL also brings back NodeBuilder and WindowBuilder to mock the
AccessibilityWindowInfo, because a real one can't be created in
the Instrumentation test.
Since the test uses both real nodes and mocked nodes, this CL uses
an updated version of NodeBuilder and WindowBuilder.

Bug: 176923471
Test: atest NavigatorTest
Test: atest NodeBuilderTest
Test: atest WindowBuilderTest
Change-Id: I6f20c6742eebfc44c5aa448817624611dbe70ebb