Make nudging up close IME and return to EditText
Make nudging up out of the IME window close the IME window and return
rotary focus to the field being edited. Note that when focus is in the
IME window, the field being edited remains focused, so returning focus
to the field being edited doesn't affect the field. It just changes the
logical focus: the view from which subsequent rotate and nudge actions
will navigate.

Also fix a few nits suggested by IntelliJ.

Test: nudge up from IME in Rotary Playground, Settings search, and
      Driver name
Test: close IME and then navigate in above apps
Bug: 174613802
Bug: 175410458
Change-Id: I2f9ac2b66183e00c6eeacdb0499c3f7084cf6e9b