Allow entering DM mode when the focus is on child
Usually entering DM mode requires that the focus view is the view that
was selected. However, in the case of Date/Time Pickers, it is possible
that we want to focus one of the child number pickers. In this CL we
relax the constraint to allow for this use case.

Also replace Utils.hasFocus, since it seems to be problematic and return
the wrong result sometimes.

Bug: 175245949
Test: manual
Change-Id: Ifad653683830b2154a7ac777cf6c329426dd44aa