Fix wrong IME issue when RotaryService got killed
When RotaryService got killed and restarts, it doesn't check whether
the IME configuration is correct, therefore the wrong IME may show
up. This CL fixes the issue.

Besides, this CL also resets to touch IME if the current IME is
rotary IME when RotaryService is destroyed.

Bug: 169423887
Bug: 177606534
Test: Switch to rotary mode, kill RotaryService manually, then test
IME. Then restart RotaryService, and test IME.

Change-Id: Ib72ebc6956f4b4b96639dfdda581016a368320e8