DO NOT MERGE Defer IME validity checks
The IMEs specified by the OEM may not be enabled yet when the
RotaryService starts up for the first time so defer checking if they're
valid until they're used to switch IMEs.

Treat disabled system IMEs as valid. This allows switching to the rotary
IME even if it's disabled for touch.

Allow switching to the rotary IME when the touch IME isn't valid and
remove validity check for touch IME since it was the current IME
regardless of its validity.

Do a full match rather than a substring match when checking if an IME is
listed in a secure settings.

Test: Use rotary and touch IMEs after running through setup wizard
Bug: 183050097
Change-Id: I62ebc66f53ea295e7af303501fa2fc749ac61301