Exclude dump code and test acceleration
Exclude dump code from test coverage reports.

Test rotation acceleration.

Fix a bug in rotation acceleration. Use the timestamp of the last
detent in the batch of events rather than the first one when
calculating the time delta for the entire batch. The Seahawk VHAL
uses zeros for the inter-detent deltas so this change will have no
effect on Seahawk.

Test: atest RotaryServiceTest
Bug: 190783937
Bug: 190782718
Change-Id: I55ceda2206bb6f0096d97d82691f42336da06fa1