Fix flaky rotary tests
Set FLAG_RETRIEVE_INTERACTIVE_WINDOWS flag so UiAutomation can
reliably access the root node in the active window.

Supress system window tests which relied on the tests not having
permission to create AccessibilityWindowInfos.

Make RotaryService not throw an exception when it has no base
context, as is the case during tests.

Test: atest CarRotaryControllerUnitTests
Bug: 194717628
Change-Id: I04b3b16e87866d6abf2f8fdc2782f10871b102c2