Fix back button event handling
When not in direct manipulation mode, inject back button events so
that the application can handle them. Returning false for captured
events doesn't allow them to propogate to the application. The code
to inject events was already present but it was only being called in
direct manipulation mode.

Test: Manual on Seahawk and emulator built from source so it includes
https: //android-review.googlesource.com/c/platform/external/qemu/+/1878757/:
  1. Follow steps in the bug
  2. Rotary Playground Sys UI Manipulation slider
  3. Adjust brightness in Settings
  4. Adjust brightness in Quick Controls
Bug: 207152818
Change-Id: Idc59880fd864264da11239567568e5b60723c00a